Complete session in verify review controller
**Why**: We were creating a profile in the verify/review#create method.
That method redirects to verify/confirmations#show which then calls
`#complete_session`. This means that the profile exists in quasi
complete state between these requests. This commit makes a small change
so the profile is created and completed in one step. Additionally, this
commit cleans up the confirmations controller specs.